Support for all XO models
Each XO model requires a OS image specifically built. The
open firmware command for flashing the OS image are identical
between models, with the exception of the XO-1. Therefore, it
could be a problem to have a script that does not consider this
fact. ie., plugin the XO4 USB stick in a XO1.75.

This patch uses a already known name convention for OS images,
detects the XO model, and runs the update command with the proper
OS image.

This is also useful for having a universal update stick. ie., one
stick that can safely upgrade all XO versions.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>