olpc.fth: check critical low and present bits
An update may still fail half-way through because of power loss.

Changes arise from review of b948250 "Check device has 50% or more
battery life".

The battery state of charge value cannot be trusted if the battery is
not present, or if the critical low state is detected.  So in this patch
OLPC shows how to check these conditions.

The firmware already has the word ?enough-power, which also checks that
external power is available.

: ?enough-power  ( )
   bat-status@                                    ( stat )
   dup  h# 10 and  0=  abort" No external power"  ( stat )
   dup  1 and  0=  abort" No battery"             ( stat )
   4 and  abort" Battery low"
;

OLPC recommends this check, as it prevents bricking in field caused by a
firmware reflash after electrostatic discharge to the battery EEPROM
through the battery terminals.

http://wiki.laptop.org/go/SPI_FLASH_Recovery/XO-1.5_Partial

However, we are aware that the Australian deployment has circumvented
this check so that chargers need not be available during firmware
reflash.  So on this patch the deployment policy is upheld, and the
check of bit 4 is omitted.